requirements.txt
# ==================================================
# VISUAL STRATEGY CREATOR - DEPENDENCIAS
# ==================================================
# Actualizado: Enero 2026
# Python 3.8+

# ==================================================
# FRAMEWORK WEB
# ==================================================
flask==3.0.0
werkzeug==3.0.1

# ==================================================
# API DE EXCHANGES Y DATOS
# ==================================================
# NOTA: Binance/Bybit NO son para backtest (solo trading en vivo)
# Para backtest usamos proveedores confiables:
yfinance>=0.2.32  # Yahoo Finance - Datos históricos confiables
requests>=2.31.0  # Para CoinGecko API y Binance API pública (gratuito, sin restricciones)

# ==================================================
# ANÁLISIS DE DATOS
# ==================================================
pandas==2.1.4
numpy==1.26.2

# ==================================================
# BACKTESTING
# ==================================================
backtesting==0.3.3

# ==================================================
# INDICADORES TÉCNICOS
# ==================================================
ta==0.11.0

# ==================================================
# VISUALIZACIÓN
# ==================================================
bokeh==3.3.2

# ==================================================
# BASE DE DATOS
# ==================================================
# SQLite viene con Python, no requiere instalación adicional

# ==================================================
# AUTENTICACIÓN
# ==================================================
python-dotenv==1.0.0
PyJWT==2.8.0

# ==================================================
# GOOGLE OAUTH
# ==================================================
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# ==================================================
# EMAIL SERVICE
# ==================================================
# smtplib viene con Python

# ==================================================
# UTILIDADES
# ==================================================
python-dateutil==2.8.2
requests==2.31.0

# ==================================================
# PRODUCCIÓN (OPCIONAL)
# ==================================================
# Para despliegue en VPS, recomendamos:
gunicorn==21.2.0  # Servidor WSGI para producción
